+++ src/renderer/editor/properties/properties.ts
@@ -8,27 +8,89 @@
 // SPDX-FileCopyrightText: Simon Schneegans <code@simonschneegans.de>
 // SPDX-License-Identifier: MIT
 
+import { EventEmitter } from 'events';
 import Handlebars from 'handlebars';
 
-export class Properties {
+import { IEditorMenuItem } from '../common/editor-menu-item';
+import { IMenu } from '../../../common';
+import { IconPicker } from './icon-picker';
+import { IconThemeRegistry } from '../../../common/icon-theme-registry';
+
+/**
+ * This class is responsible for displaying the properties of the currently edited menu
+ * item. It emits events when the user changes one of the properties of the currently
+ * edited menu item. The new values are stored in the menu item object before the
+ * corresponding event is emitted, so you can access the new values by accessing the
+ * properties of the menu item object. It is an event emitter which emits the following
+ * events:
+ *
+ * @fires changed-name - When the user changed the name of the current menu item.
+ * @fires changed-icon - When the user changed the icon of the current menu item.
+ */
+export class Properties extends EventEmitter {
   // The container is the HTML element which contains the currently edited menu's
   // properties. It is created in the constructor and returned by the getContainer()
   // method.
   private container: HTMLElement = null;
+
+  private iconPicker: IconPicker = null;
+
+  private baseSettings: HTMLElement = null;
+  private nameInput: HTMLInputElement = null;
+  private iconButton: HTMLButtonElement = null;
+
+  private activeItem: IEditorMenuItem = null;
 
   /**
    * This constructor creates the HTML elements for the menu properties view and wires up
    * all the functionality.
    */
   constructor() {
+    super();
+
     const template = Handlebars.compile(require('./templates/properties.hbs').default);
 
     const div = document.createElement('div');
-    div.innerHTML = template({
-      areaId: 'kando-menu-properties-area',
+    div.innerHTML = template({});
+
+    this.nameInput = div.querySelector('#kando-menu-properties-name') as HTMLInputElement;
+    this.nameInput.addEventListener('input', () => {
+      if (this.activeItem) {
+        this.activeItem.name = this.nameInput.value;
+        this.emit('changed-name');
+      }
+    });
+
+    this.iconButton = div.querySelector(
+      '#kando-menu-properties-icon-button'
+    ) as HTMLButtonElement;
+    this.iconButton.addEventListener('click', () => {
+      this.iconPicker.show(this.activeItem.icon, this.activeItem.iconTheme);
+      this.baseSettings.classList.add('hidden');
     });
 
     this.container = div.firstElementChild as HTMLElement;
+
+    this.baseSettings = div.querySelector('#kando-menu-properties-base-settings');
+
+    this.iconPicker = new IconPicker(
+      div.querySelector('#kando-menu-properties-icon-picker')
+    );
+    this.iconPicker.on('select', (icon, theme) => {
+      if (this.activeItem) {
+        this.activeItem.icon = icon;
+        this.activeItem.iconTheme = theme;
+
+        this.iconButton.innerHTML = IconThemeRegistry.getInstance()
+          .getTheme(theme)
+          .createDiv(icon).outerHTML;
+
+        this.emit('changed-icon');
+      }
+    });
+    this.iconPicker.on('close', () => {
+      this.baseSettings.classList.remove('hidden');
+    });
   }
 
   /** This method returns the container of the menu preview. */
@@ -42,5 +104,22 @@ export class Properties {
 
   public hide() {
     this.container.classList.remove('visible');
+  }
+
+  public setMenu(menu: IMenu) {
+    this.setItem(menu.nodes);
+  }
+
+  public setItem(item: IEditorMenuItem) {
+    if (this.activeItem !== item) {
+      this.activeItem = item;
+      this.nameInput.value = item.name;
+      this.iconButton.innerHTML = IconThemeRegistry.getInstance()
+        .getTheme(item.iconTheme)
+        .createDiv(item.icon).outerHTML;
+
+      this.baseSettings.classList.remove('hidden');
+      this.iconPicker.hide();
+    }
   }
 }